ecomerce/requirements.txt
django
psycopg2-binary
djangorestframework
tzdata
djangorestframework_simplejwt